feat: Add user detail page and update user item component
The code changes include adding a user detail page and updating the user item component to include a link to the user detail page. This enhances the user experience by allowing users to view more information about a specific user when clicking on the "More" button in the user item component.